+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * e83d858 - 2016-10-13: new: v8.5.4 fix: Removed old reference to theTVDB API v1 
+- * 652d3b5 - 2016-10-13: new: v8.5.5 fix: Post-Processing handles deleted episode exceptions properly now 
+- * 1c30b91 - 2016-10-13: new: v8.5.4 fix: Removed old reference to theTVDB API v1 
 - * a337fb8 - 2016-10-13: new: v8.5.3 
 - * c519929 - 2016-10-13: fix: RevConflict issue during updating of shows and imdb info was caused by mistakenly storing previous rev info to class object 
 - * 1c7a1ba - 2016-10-13: fix: Changed web session handling of error messages to send to debugger instead of creating a error